requirements.txt
qrcode[pil]
Pillow
customtkinter
pandas
openpyxl
matplotlib
reportlab
pyinstaller
